Simplify algorithm to create contour from coordinates
